requirements.txt
Flask==3.1.3
Flask-Caching==2.3.1
flask-cors==6.0.2
python-dotenv==1.2.2
requests==2.33.1
pytest==9.0.3
